% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotAccountingType}
\alias{GetAccountSnapshotAccountingType}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotAccountingType()
}
\description{
Get Account snapshot field
}
