% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotBalanceCurrency}
\alias{GetAccountSnapshotBalanceCurrency}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotBalanceCurrency()
}
\description{
Get Account snapshot field
}
