% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotBalanceToReportConversionRate}
\alias{GetAccountSnapshotBalanceToReportConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotBalanceToReportConversionRate()
}
\description{
Get Account snapshot field
}
