% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotBalanceToUsdConversionRate}
\alias{GetAccountSnapshotBalanceToUsdConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotBalanceToUsdConversionRate()
}
\description{
Get Account snapshot field
}
