% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotProfitToReportConversionRate}
\alias{GetAccountSnapshotProfitToReportConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotProfitToReportConversionRate()
}
\description{
Get Account snapshot field
}
