% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotProfitToUsdConversionRate}
\alias{GetAccountSnapshotProfitToUsdConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotProfitToUsdConversionRate()
}
\description{
Get Account snapshot field
}
