% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotReportCurrency}
\alias{GetAccountSnapshotReportCurrency}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotReportCurrency()
}
\description{
Get Account snapshot field
}
