% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotReportToBalanceConversionRate}
\alias{GetAccountSnapshotReportToBalanceConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotReportToBalanceConversionRate()
}
\description{
Get Account snapshot field
}
