% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotReportToProfitConversionRate}
\alias{GetAccountSnapshotReportToProfitConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotReportToProfitConversionRate()
}
\description{
Get Account snapshot field
}
