% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotTotalProfitLoss}
\alias{GetAccountSnapshotTotalProfitLoss}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotTotalProfitLoss()
}
\description{
Get Account snapshot field
}
