% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{GetAccountSnapshotUsdToBalanceConversionRate}
\alias{GetAccountSnapshotUsdToBalanceConversionRate}
\title{Get Account snapshot field}
\usage{
GetAccountSnapshotUsdToBalanceConversionRate()
}
\description{
Get Account snapshot field
}
