% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snaphots.R
\name{GetSnapshotCurrencyToReportConversionRate}
\alias{GetSnapshotCurrencyToReportConversionRate}
\title{Get Snapshot field}
\usage{
GetSnapshotCurrencyToReportConversionRate()
}
\description{
Get Snapshot field
}
