% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snaphots.R
\name{GetSnapshotReportToCurrencyConversionRate}
\alias{GetSnapshotReportToCurrencyConversionRate}
\title{Get Snapshot field}
\usage{
GetSnapshotReportToCurrencyConversionRate()
}
\description{
Get Snapshot field
}
