% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeBalanceToReportConversionRate}
\alias{GetTradeBalanceToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeBalanceToReportConversionRate()
}
\description{
Get Trade report field
}
