% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeDstAssetToReportConversionRate}
\alias{GetTradeDstAssetToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeDstAssetToReportConversionRate()
}
\description{
Get Trade report field
}
