% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeMarginCurrencyToReportConversionRate}
\alias{GetTradeMarginCurrencyToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeMarginCurrencyToReportConversionRate()
}
\description{
Get Trade report field
}
