% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeProfitCurrencyToReportConversionRate}
\alias{GetTradeProfitCurrencyToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeProfitCurrencyToReportConversionRate()
}
\description{
Get Trade report field
}
