% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeProfitCurrencyToUsdConversionRate}
\alias{GetTradeProfitCurrencyToUsdConversionRate}
\title{Get Trade report field}
\usage{
GetTradeProfitCurrencyToUsdConversionRate()
}
\description{
Get Trade report field
}
