% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeProfitToReportConversionRate}
\alias{GetTradeProfitToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeProfitToReportConversionRate()
}
\description{
Get Trade report field
}
