% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeSrcAssetToReportConversionRate}
\alias{GetTradeSrcAssetToReportConversionRate}
\title{Get Trade report field}
\usage{
GetTradeSrcAssetToReportConversionRate()
}
\description{
Get Trade report field
}
