% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeUsdToMarginCurrencyConversionRate}
\alias{GetTradeUsdToMarginCurrencyConversionRate}
\title{Get Trade report field}
\usage{
GetTradeUsdToMarginCurrencyConversionRate()
}
\description{
Get Trade report field
}
