% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{GetTradeUsdToProfitCurrencyConversionRate}
\alias{GetTradeUsdToProfitCurrencyConversionRate}
\title{Get Trade report field}
\usage{
GetTradeUsdToProfitCurrencyConversionRate()
}
\description{
Get Trade report field
}
