% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{ttmConnect}
\alias{ttmConnect}
\title{Connects to a TT server}
\usage{
ttmConnect(address = "", login = "", password = "")
}
\arguments{
\item{address}{Url of the server}

\item{login}{account you login}

\item{password}{password of the account}
}
\description{
Connects to a TT server
}
