% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{ttmCreateNewOrder}
\alias{ttmCreateNewOrder}
\title{Create the Orders as requested}
\usage{
ttmCreateNewOrder(login, orderType, orderSide, symbol, amount, stopPrice,
  price, stopLoss, takeProfit, UserComment, ManagerComment, Expiration)
}
\description{
Create the Orders as requested
}
\examples{
ttmCreateNewOrder()

}
