% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmCreateNewSymbol}
\alias{ttmCreateNewSymbol}
\title{Create new Symbol}
\usage{
ttmCreateNewSymbol(
  symbol,
  security,
  isin,
  alias,
  marginCurrency,
  profitCurrency,
  precision,
  contractSize,
  marginFactorFractional = 1,
  swapEnabled = TRUE,
  marginMode = "CFD",
  profitMode = "CFD",
  marginHedged = 0.5,
  swapType = "PercentPerYear",
  swapSizeLong = 0,
  swapSizeShort = 0,
  description = ""
)
}
\arguments{
\item{symbol}{a character. Symbol}

\item{security}{a character. Security}

\item{isin}{a character. ISIN}

\item{alias}{a character. Alias}

\item{marginCurrency}{a character. MarginCurrency}

\item{profitCurrency}{a character. ProfitCurrency}

\item{precision}{a numeric. Precision}

\item{contractSize}{a numeric. ContractSize}

\item{marginFactorFractional}{a numeric. MarginFactorFractional}

\item{swapEnabled}{a boolean. SwapEnabled}

\item{marginMode}{a character. MarginMode}

\item{profitMode}{a character. ProfitMode}

\item{marginHedged}{a numeric.  MarginHedged}

\item{swapType}{a character. SwapType. For TTS only "Points" or "PercentPerYear"}

\item{swapSizeLong}{a numeric. SwapSizeLong}

\item{swapSizeShort}{a numeric SwapSizeShort}

\item{description}{a character. SwapSizeShort}
}
\description{
Create new Symbol
}
