% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmDeleteQuotes}
\alias{ttmDeleteQuotes}
\title{Delete Quotes}
\usage{
ttmDeleteQuotes(symbol, from, to, periodicityLevel)
}
\arguments{
\item{symbol}{a character vector. Symbol names}

\item{from}{a DateTime object. Start time to delete}

\item{to}{a DateTime object. End time to delete}

\item{periodicityLevel}{a character. Periodicity Level. From this ("TicksLevel2", "Ticks" ,"M1", "H1", "VWAP")}
}
\description{
Delete Quotes
}
\examples{
ttmExportQuotes(c("EURUSD"), ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), ISOdatetime(2019,06,06,0,00,00, tz ="GMT"), "M1", "C:/Quotes", TRUE)

}
