% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmDeleteSymbolHistory}
\alias{ttmDeleteSymbolHistory}
\title{Delete Quotes for all Periodicity}
\usage{
ttmDeleteSymbolHistory(symbol, from, to)
}
\arguments{
\item{symbol}{a character vector. Symbol names}

\item{from}{a DateTime object. Start time to delete}

\item{to}{a DateTime object. End time to delete}
}
\description{
Delete Quotes for all Periodicity
}
\examples{
ttmDeleteSymbolHistory(c("EURUSD"), ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), ISOdatetime(2019,06,06,0,00,00, tz ="GMT"))

}
