% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accounts.R
\name{ttmGetAccountByLogin}
\alias{ttmGetAccountByLogin}
\title{Gets the Accounts by login}
\usage{
ttmGetAccountByLogin(accountId, fromArchieve = FALSE)
}
\description{
Gets the Accounts by login
}
\examples{
GetAccountByLogin(100040)
}
