% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccountSnaphots.R
\name{ttmGetAccountSnaphots}
\alias{ttmGetAccountSnaphots}
\title{Gets the Account snaphots as requested}
\usage{
ttmGetAccountSnaphots(accId, from = ISOdatetime(1970, 1, 1, 0, 0, 0, tz =
  "GMT"), to = ISOdatetime(2017, 8, 1, 0, 0, 0, tz = "GMT"))
}
\arguments{
\item{accId}{a numeric vector. Accounts ids.}

\item{from}{a POSIXct object. Start time. By default, from = ISOdatetime(1970,01,01,0,00,00, tz ="GMT").}

\item{to}{a POSIXct object. End time. By default, to = ISOdatetime(2017,08,01,0,00,00, tz ="GMT").}
}
\description{
Gets the Account snaphots as requested
}
\examples{
ttmGetAccountSnaphots(c(100181,100182,100183), ISOdatetime(1970,01,01,0,00,00, tz ="GMT"), ISOdatetime(2017,08,01,0,00,00, tz ="GMT"))

}
