% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskInfo.R
\name{ttmGetAllCorporateActionFeedTasks}
\alias{ttmGetAllCorporateActionFeedTasks}
\title{Gets the Task info (only CorporateActionFeedTasks)}
\usage{
ttmGetAllCorporateActionFeedTasks()
}
\description{
Gets the Task info (only CorporateActionFeedTasks)
}
\examples{
ttmGetAllCorporateActionFeedTasks()

}
