% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{ttmGetAllGroupsSecurityInfo}
\alias{ttmGetAllGroupsSecurityInfo}
\title{Gets the groups as requested}
\usage{
ttmGetAllGroupsSecurityInfo()
}
\description{
Gets the groups as requested
}
\examples{
ttmGetAllGroupsSecurityInfo()
}
