% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetPositions.R
\name{ttmGetAllPositions}
\alias{ttmGetAllPositions}
\title{Gets All Net Positions as requested}
\usage{
ttmGetAllPositions()
}
\description{
Gets All Net Positions as requested
}
\examples{
ttmGetAllPositions()

}
