% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmGetAllSymbols}
\alias{ttmGetAllSymbols}
\title{#' Create New Symbol (For Test Only)
ttmCreateNewSymbol <- function(symbolName = "",security = "",marginCurrency = "",profitCurrency = "", precision = 5, contractSize = 1) {
  hResult = rClr::clrCallStatic('rTTManApi.rTTManApiHost', 'CreateSymbol', symbolName, security, marginCurrency, profitCurrency, precision, contractSize)
  return(hResult)
}
Gets the symbols as requested}
\usage{
ttmGetAllSymbols()
}
\description{
#' Create New Symbol (For Test Only)
ttmCreateNewSymbol <- function(symbolName = "",security = "",marginCurrency = "",profitCurrency = "", precision = 5, contractSize = 1) {
  hResult = rClr::clrCallStatic('rTTManApi.rTTManApiHost', 'CreateSymbol', symbolName, security, marginCurrency, profitCurrency, precision, contractSize)
  return(hResult)
}
Gets the symbols as requested
}
\examples{
ttmGetAllSymbols()

}
