% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bars.R
\name{ttmGetBarHistory}
\alias{ttmGetBarHistory}
\title{Get Bars History}
\usage{
ttmGetBarHistory(symbol, periodicity, barType, endTime, count)
}
\arguments{
\item{symbol}{a character. Symbol name}

\item{periodicity}{a character. Bar periodicity}

\item{barType}{a character. Ask or Bid}

\item{endTime}{a DateTime object. End time}

\item{count}{a numeric. Integer ticks count (Can be positive or negative)}
}
\description{
Get Bars History
}
\examples{
ttmGetTicks("EURUSD", ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), -100)

}
