% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{ttmGetOrders}
\alias{ttmGetOrders}
\title{Gets the Orders as requested with login}
\usage{
ttmGetOrders(accountId)
}
\arguments{
\item{accountId}{a numeric. Account id}
}
\description{
Gets the Orders as requested with login
}
