% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduleInfo.R
\name{ttmGetScheduleInfo}
\alias{ttmGetScheduleInfo}
\title{Gets the Schedule OffTimes requested}
\usage{
ttmGetScheduleInfo()
}
\description{
Gets the Schedule OffTimes requested
}
\examples{
ttmGetScheduleInfo()

}
