% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmGetTickHistoryCache}
\alias{ttmGetTickHistoryCache}
\title{Get Ticks History Cache}
\usage{
ttmGetTickHistoryCache(symbol, endTime, count)
}
\arguments{
\item{symbol}{a character. Symbol name}

\item{endTime}{a DateTime object. End time}

\item{count}{a numeric. Integer ticks count (Can be positive (to future) or negative (to past))}
}
\description{
Get Ticks History Cache
}
\examples{
ttmGetTickHistoryCache("EURUSD", ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), -100)

}
