% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmGetTicks}
\alias{ttmGetTicks}
\title{Get Quotes}
\usage{
ttmGetTicks(symbol, endTime, count)
}
\arguments{
\item{symbol}{a character. Symbol name}

\item{count}{a numeric. Integer ticks count (Can be positive or negative)}

\item{to}{a DateTime object. End time}
}
\description{
Get Quotes
}
\examples{
ttmGetTicks("EURUSD", ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), -100)

}
