% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmGetTicksHistory}
\alias{ttmGetTicksHistory}
\title{Get Ticks History}
\usage{
ttmGetTicksHistory(symbol, endTime, count)
}
\arguments{
\item{symbol}{a character. Symbol name}

\item{endTime}{a DateTime object. End time}

\item{count}{a numeric. Integer ticks count (Can be positive (to past) or negative (to future))}
}
\description{
Get Ticks History
}
\examples{
ttmGetTicksHistory("EURUSD", ISOdatetime(2019,06,05,0,00,00, tz ="GMT"), -100)

}
