% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmInsertLevel2Ticks}
\alias{ttmInsertLevel2Ticks}
\title{Upload symbol ticks L2}
\usage{
ttmInsertLevel2Ticks(symbol, timestamps, bidPrices, bidVolumes, askPrices,
  askVolumes, depth)
}
\arguments{
\item{symbol}{Symbol}

\item{timestamps}{Array of datetimes}

\item{bidPrices}{Array of bid prices}

\item{bidVolumes}{Array of bid volumes}

\item{askPrices}{Array of ask prices}

\item{askVolumes}{Array of ask volumes}

\item{depth}{Array of depth}
}
\description{
Upload symbol ticks L2
}
