% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmInsertTicks}
\alias{ttmInsertTicks}
\title{Upload symbol ticks}
\usage{
ttmInsertTicks(symbol, timestamps, bidPrices, bidVolumes, askPrices, askVolumes)
}
\arguments{
\item{symbol}{Symbol}

\item{timestamps}{Array of datetimes}

\item{bidPrices}{Array of bid prices}

\item{bidVolumes}{Array of bid volumes}

\item{askPrices}{Array of ask prices}

\item{askVolumes}{Array of ask volumes}
}
\description{
Upload symbol ticks
}
