% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmModifySymbol}
\alias{ttmModifySymbol}
\title{Modify Symbol}
\usage{
ttmModifySymbol(symbol, isin)
}
\arguments{
\item{symbol}{a character. Symbol}

\item{isin}{a character. New ISIN value}
}
\description{
Modify Symbol
}
