% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmModifySymbolExtendedNameDescription}
\alias{ttmModifySymbolExtendedNameDescription}
\title{Modify Symbol ExtendedName and Description}
\usage{
ttmModifySymbolExtendedNameDescription(symbol, newExtendedName, newDescription)
}
\arguments{
\item{newExtendedName}{a character. New Extended Name value}

\item{newDescription}{a character. New Description value}

\item{symbolId}{a character. Symbol}
}
\description{
Modify Symbol ExtendedName and Description
}
