% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect.R
\name{ttmConnect}
\alias{ttmConnect}
\title{Connects to a TT server}
\usage{
ttmConnect(address = "", login = "", password = "", file = "")
}
\arguments{
\item{address}{Url of the server}

\item{login}{account you login}

\item{password}{password of the account}

\item{file}{when not empty, credentials will be read from it}
}
\description{
Connects to a TT server
}
\examples{
ttmConnect("tpstsoft.eu", "123456","123456", "fileName.properties")

}
