% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{ttmGetTradeReports}
\alias{ttmGetTradeReports}
\title{Gets the Trade reports as requested}
\usage{
ttmGetTradeReports(accId, from = as.POSIXct(0, origin = ISOdatetime(1970, 1,
  1, 3, 0, 0)), to = Sys.time(), skipCancelled = TRUE)
}
\description{
Gets the Trade reports as requested
}
