% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TradeHistory.R
\name{ttmGetTradeReports}
\alias{ttmGetTradeReports}
\title{Gets the Trade reports as requested}
\usage{
ttmGetTradeReports(accId, from = ISOdatetime(1970, 1, 1, 0, 0, 0, tz = "GMT"),
  to = ISOdatetime(2017, 8, 1, 0, 0, 0, tz = "GMT"), skipCancelled = TRUE)
}
\arguments{
\item{accId}{a numeric vector. Accounts ids.}

\item{from}{a POSIXct object. Start time. By default, from = ISOdatetime(1970,01,01,0,00,00, tz ="GMT").}

\item{to}{a POSIXct object. End time. By default, to = ISOdatetime(2017,08,01,0,00,00, tz ="GMT").}

\item{skipCancelled}{a logical. If TRUE (default), cancelled orders are not displayed.}
}
\description{
Gets the Trade reports as requested
}
