% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmInsertLevel2Ticks}
\alias{ttmInsertLevel2Ticks}
\title{Upload symbol ticks L2}
\usage{
ttmInsertLevel2Ticks(symbol, timestamps, bidPrices, bidVolumes, askPrices,
  askVolumes, depth)
}
\arguments{
\item{symbol}{Symbol}

\item{timestamps}{a DateTime array. Datetimes}

\item{bidPrices}{a double array. Bid prices}

\item{bidVolumes}{a double array. Bid volumes}

\item{askPrices}{a double array. Ask prices}

\item{askVolumes}{a double array. Ask volumes}

\item{depth}{a double array. Depths}
}
\description{
Upload symbol ticks L2
}
