% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmInsertTicks}
\alias{ttmInsertTicks}
\title{Upload symbol ticks}
\usage{
ttmInsertTicks(symbol, timestamps, bidPrices, bidVolumes, askPrices, askVolumes)
}
\arguments{
\item{symbol}{Symbol}

\item{timestamps}{a DateTime array. Datetimes}

\item{bidPrices}{a a double array. Bid prices}

\item{bidVolumes}{a double array. Bid volumes}

\item{askPrices}{a double array. Ask prices}

\item{askVolumes}{a double array. Ask volumes}
}
\description{
Upload symbol ticks
}
\examples{
ttmInsertTicks("EURUSD", c(ISOdatetime(2017,11,01,0,00,00, tz ="GMT"), ISOdatetime(2017,11,01,0,00,02, tz ="GMT")), c(1.20, 1.21), c(100, 1000), c(1.21,1.22), c(1000, 1000))

}
