% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmModifySymbolSwap}
\alias{ttmModifySymbolSwap}
\title{Change symbol swap}
\usage{
ttmModifySymbolSwap(
  symbol = "",
  swapSizeShort = "",
  swapSizeLong = "",
  swapType = ""
)
}
\arguments{
\item{symbol}{a character. Symbol}

\item{swapSizeShort}{a double. SwapSizeShort}

\item{swapSizeLong}{a double. SwapSizeLong}

\item{swapType}{a character. SwapType. For TTS only "Points" or "PercentPerYear"}
}
\description{
Change symbol swap
}
\examples{
ttmModifySymbolSwap("BTCBYN", 10, 10, "Points")

}
