% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbols.R
\name{ttmModifySymbolSwap}
\alias{ttmModifySymbolSwap}
\title{Change symbol swap}
\usage{
ttmModifySymbolSwap(symbol = "", swapSizeShort = "", swapSizeLong = "")
}
\arguments{
\item{symbol}{Name of symbol}

\item{swapSizeShort}{swapSizeShort}

\item{swapSizeLong}{swapSizeLong}
}
\description{
Change symbol swap
}
