% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmUpstream}
\alias{ttmUpstream}
\title{Upstream symbol ticks}
\usage{
ttmUpstream(symbol, from, to)
}
\arguments{
\item{symbol}{a character. Symbol}

\item{from}{a DateTime object. Start time to upstream}

\item{to}{a DateTime object. End time to upstream}
}
\description{
Upstream symbol ticks
}
