% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ticks.R
\name{ttmUpstreamAsync}
\alias{ttmUpstreamAsync}
\title{UpstreamAsync symbol ticks}
\usage{
ttmUpstreamAsync(symbol, from, to, upstreamType)
}
\arguments{
\item{symbol}{a character. Symbol}

\item{from}{a DateTime object. Start time to upstream}

\item{to}{a DateTime object. End time to upstream}

\item{upstreamType}{a numeric. UpstreamType Enum Code}
}
\description{
UpstreamAsync symbol ticks
}
\examples{
 ttmUpstream("EURUSD", ISOdatetime(2017,08,01,0,00,00, tz ="GMT"), ISOdatetime(2017,08,02,0,00,00, tz ="GMT"), 1)

}
